requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
ecdsa==0.19.1
executing==2.2.0
Faker==37.1.0
fastapi==0.115.12
fastembed==0.6.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.4
onnxruntime==1.21.1
openai==1.75.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
portalocker==2.10.1
prompt_toolkit==3.0.51
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.13.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tinycss2==1.4.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
